+++ sync-for-payroll/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.1.1] sync-for-payroll
 ### Releases
 - [PyPI v0.1.1] https://pypi.org/project/codat-sync-for-payroll/0.1.1 - sync-for-payroll
+
+## 2023-09-05 00:09:58
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.2] sync-for-payroll
+### Releases
+- [PyPI v0.1.2] https://pypi.org/project/codat-sync-for-payroll/0.1.2 - sync-for-payroll